refactor: 승부 결정 로직을 MatchResult가 담당
